made changes as pr new folder structure
